gtk2: don't use one-line "if return"
It is not gdb friendly. We jump to the end of function by "next"
command when the "if" condition is true. If the "if" condition is
unexpected, we want to confirm the information in the local block
scope. But if we jump to the end of function, we can't confirm the
information. Because we are already in outside of the local block.