gdk3-gi test: use assert_nil instead of assert_nothing_raised
We should minimize assert_nothing_raised use because it doesn't check
anything. If any exception is raised, test-unit report it. We can
notify it without assert_nothing_raised.

assert_nothing_raised should be used only for checking "the most
important thing in the test is that the method call is valid". Return
value check is better in most cases.