evaluating local_gemfile using Bundler DSL
Within the scope of a Gemfile evaluated by bundler, the object
referenced as 'self' may usually be an object of a type Bundler::Dsl

Using the eval_gemfile method on the the Bundler::Dsl instance when
evaluating an existing Gemfile.local, Bundler will then use a
Bundler::Dsl syntax for the file. This may also ensure that the filename
is added to any Gemfile state data for the Bundler::Dsl instance

With this patch, a call to the 'gem' method in a Gemfile.local may
then be interpreted by bundler as it being a reference to the 'gem'
method on the Bundler::Dsl instance. This may differ to how a call to
the 'gem' method would be evaluated in ordinary eval.

e.g using Gemfile.local to ensure that an interactive Ruby interpreter
of any local preference is available under 'bundle exec'

a Gemfile.local in excerpt, for IRB:
~~~~
gem 'irb'
~~~~

alternately, for Pry:
~~~~
gem 'pry'
~~~~

(cherry picked from commit d5885c706d738ec79dfdcf151e84c01690600084)