glib2: fix a crash bug that is related to closure and GC
GRClosure is a GClosure for Ruby. It holds Ruby's Proc and TypedData
for GRClosure to mark held Ruby's Proc. The TypedData for GRClosure is
referred from a related Ruby object. For example,
"object" is the related Ruby object for "object.signal_connect(...) {}".

gr_closure_holder_free() is called when the TypedData for GRClosure is
GC-ed (the related Ruby object is GC-ed). It unrefers the related
GRClosure. If the reference count of the unref-ed GRClosure becomes 0,
there is no problem. Because rclosure_finalize() is called
immediately. rclosure_finalize() removes all related references to the
GC-ed GRClosure. And it also sets
RTYPEDDATA_DATA(GRClosure::rb_holder) to NULL.

This (RTYPEDDATA_DATA(GRClosure::rb_holder) = NULL) is a problem when
the reference count of the unref-ed GRClosure doesn't become 0. In the
case, rclosure_finalize() isn't called immediately. But
GRClosure::rb_holder's (TypedData of GRClosure) address may be reused
because the GRClosure::rb_holder is GC-ed. If an object is allocated
to the address that was used by GC-ed GRClosure::rb_holder,
RTYPEDDATA_DATA(GRClosure::rb_holder) = NULL in rclosure_finalize()
breaks the object that was newly allocated to the same address.

We can fix this by calling g_closure_invalidate() in
gr_closure_holder_free() explicitly and use invalidate notification
instead of finalize notification. rclosure_invalidate() must be called
immediately when the GRClosure (TypedData of it) is GC-ed.