glib2: remove rbg_define_method()
GitHub: GH-1660

Because ANYARG isn't compatible with C23.

This adds rbg_define_setter_alias_if_need() and use rb_define_method()
and it in RG_DEF_METHOD().

This is a backward incompatible change but this will not be a problem
because RG_DEF_METHOD() API isn't changed. (ABI is incompatible.)

Reported by Mamoru TASAKA. Thanks!!!